requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
pymupdf
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
pandas
openai
langchain-groq
duckduckgo-search
streamlit
